backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchtext==0.16.1
pandas==2.0.3
numpy==1.24.3
langid==1.1.6
google-generativeai==0.8.3
python-dotenv==1.0.0
